Add CefSharp.Wpf integration
This currently just puts up Hello World html in a transparent window
with a partially transparent square.

The CefSharp integration requires a number of files to be copied
directly into the ACT folder so that it can load them.  These
are all the files listed here:

https://github.com/cefsharp/CefSharp/wiki/Getting-Started#copying-dependencies

If those are not there, then the CefSharp.Wpf.dll will fail to
load with mysterious error messages.  (Sorry.)

There's a terrible hack in BrowserWindow.WindowActivated, as the
web content would usually (but not always!) not appear and only the
background color would appear. Resizing the window would fix this,
and random StackOverflow answers appeared to confirm that other
folks were suffering from this as well.

It's probably just my absolute and total lack of understanding of
Wpf and Forms, but I'll leave it for now because it's working.