Switch to jquery for resizing/dragging
resize:both in Chrome has a bug (or, arguably is following the spec with
weasel words) where as soon as you start resizing an element for the
first time, its minimum size seems to be capped to that starting size.
This makes it impossible to add resize css onto an element and shrink
it.  Sad trombone noise.

Enter jquery for resizing support.  Since jquery can also do dragging,
scrap Draggabilly and use jquery for that as well.

jquery usage kept to a minimum for now other than interfacing with these
ui features, just in case it needs to get ripped out later.

Also, to squelch the error about loading images, add the icon image in.
Other images might be needed for other features, but don't seem to be
needed now, so haven't been added.