Add TP viewer window
Sadly, there's no way to figure out where this is, so it has to be
positioned manually.  Also, it can't be scaled currently either.

Because there's no way to figure out ordering from the player list,
players have to manually echo who is in their party, via a macro
like: /echo cactbot:tp:1:<1> (repeating for all 8 members).

This will fail in a dungeon when two people have the same name,
so let's hope that doesn't happen.