Auto-detect player name for TP window
This prevents the unknown "0 tp" background color from obscuring
the actual TP bar for the player before the macro has hit.  Now
the tp macro doesn't need to be hit in order for the player's
tp bar in the ui to show up.

Alternatively, the first ui bar could be hidden entirely, but then
it would be a different color from the other bars.