Add gathering window
This window uses the Garland Bell node info and utility function
for converting local time to Eorzean time and back.  The author
kindly said that it was free to use in whatever way, and so is
included here.

When the player is a miner or a botanist, the current list of
nodes will appear.  The next node or any open node will list
all of the items and the slot, along with the location
information.