Add a JS event dispatch method and push to it every 16ms
Instead of making methods to build Json, then dispatching it to a fixed
event, we have a method that dispatches to a variable event any type which
it will serialize into json and place in the the event's detail field.

The OverlayPlugin calls Update() every 1s which is not enough for something
that is reacting to logs. So add our own timer that ticks more frequently.