Add JSEvents class where all events details are defined, and onLogEvent
The onLogEvent fires when log messages have happened. Its details member is |logs| which
is a list of strings for each log message.

The |logs| field is now removed from onOverlayDataUpdate, and onLogEvent is only fired
when there are logs present, not when there are 0 logs since the last timer tick.