Add C# public events for data changes
This is to enable a future patch where some other class wants
to listen in on when things have changed without tracking
that state themselves.

Also refactor OnCombatStarted/Ended to just be a single message
with a boolean for simplicity.  Removed the wipe message here,
as that will be attached to OnFightStarted/OnFightEnded for
"known" fights with phases and names.  The combat message is
just to know whether or not ACT thinks you're in combat or not.