Add onTargetChangedEvent with info about the target.
Notes: The distance in the target struct is not exactly accurate,
it can report 25 while 25yd abilities show that they can not be
used, when you are at about 25.5 or so and up (since there's no
floats available I can only guess here).

As such, you would have to subtract 1 to be sure you can use an
ability for the given range at that time, which is super sad. Using
the position of the player and target doesn't work because the
distance for abilities takes into account the target's bounds,
but I was not able to find the bounds available in the target
struct so far.