More thread fixes for fast/slow timers
The timer will keep ticking (on multiple threads) if a task runs
long. This makes the code require to be threadsafe. Instead make
the timer not AutoReset, and call Start() from the timer callback.
This lets us run the timer slower when the game isn't found.

The fast and slow timers are on different threads, so anything they
call in common must be threadsafe. They were both using DispatchToJS
which isn't. Instead, move all the code to the fast timer, since
FindProcess() is very cheap in the common case anyway, and we can
now slow down the fast timer when the process isn't found.

Use a semaphore to alert the timer to reset NotifyState instead
of changing NotifyState from a non-timer thread which would be a
race.