Add form controls for dps/log updates
To cut down on extra cpu usage, allow plugins to opt out
of dps and log updates and specify the rate for dps
updates.  Setting the rate to zero for dps updates stops
all dps updates.  Boss start/end messages and phase updates
will still be sent, but the dps information in them will
be null.