Improvement to redmage bars
Move all config globals to the top of the file.
Use a timer-bar instead of a resource-bar for the cast
bar so that it can update at the rate of the plugin independently
of what the C# reports to JS in case it reports at a lower rate
or just when the cast starts/ends if I can figure out how to do
that reliably.