Smooth out the progress of cast bars
Because the game ticks at a different rate than we sample, there is jitter
in the progress value we see each sample, and if we use that directly as the
progress visually, the cast bar jumps around a bit as it goes. To remove that
we compare the start time of the current cast vs the last one read, and if the
difference is within the length of the plugin update interval, we consider it
the same spell, and determine the progress based on Now and the saved start
time, giving a consistent growth of progress relative to the time at which
the plugin is updating.