Add tank hp / generalize tp bar layout
It's useful for a tank to know their exact hp instead of just knowing
that they're hurt, so grow tank hp bar to include the exact number.

Then, instead of specifying an absolute position for the mp/tp bar,
just specify a margin and push everything downwards.

RDM extra bars still specified absolutely, but they're not tanks, so
everything a-ok in the layout department for now.