Drop the dependency on ACT_EnmityPlugin
Since the project is not being maintained, rather than taking over ownership
of the whole thing, just reproduce the pieces we are using here, which is quite
small already since large parts of it were already reproduced.

Adds an enum for jobs and entity types, and an EntityData structure with all
the fields we read.
Adds GetEntityData that reads from a pointer into an EntityData structure.
Uses this to implement GetTargetData() and GetSelfData() to report to JS, and
convert the various uses of Combatant before to EntityData spellings instead.