Timerbar improvements
When the bar becomes full and isn't hidden, leave it in the final state rather than
reverting back to an initial empty state.

Don't reverse the direction of the bar every time layout() happens.

Add a bit more padding to the left/right of the bar based on the font size so that
text doesn't clip.