Load timeline for the zone when zone changes from the data files
By default the filename (before the extension) is the regex used to decide to
use the file or not. But we add a  command to the file format
that can override to provide a regex for matching against the current zone.

This allows the file to work for both 'Unknown (xyz)' and 'The real zone name'
when the FFXIV plugin isn't updated for a zone name yet.