Fix entity job lookup
Job is only a byte and (apparently) they stick beast gauge numbers in the
next byte.  Without this fix, as soon as war had non-zero beast gauge it
suddenly would decide that it was some ridiculous job in the 8000's,
which would destroy all the ui confusingly.

It seems a miracle that this didn't break anything before now.