If only one window param is specified it should be centered around the sync event
According to https://github.com/anoyetta/act_timeline/blob/origin/doc/TimelineSyntax.md it seems
that if window is specified, the number given is the amount of time to look for the sync event
with half of the time before and half after the time of the sync event. The default is 5, and
other numbers behave the same way.