Add Storm's Eye countdown ui
This adds a timerbox with a dynamically computed threshold such that the
box grows large as soon as you need to complete a storm's eye combo.
Often this is after finishing a combo, but if you have time to storm's
path -> storm's eye (but just exactly) and do a fell cleave or two
after your first heavy swing it will grow big to tell you to convert
your current combo into a storm's eye.  It knows enough about combos
to not have you break a butcher's block combo and just assumes that
the storm's eye is even further out.

Some adjustments to timer box were made, in particular to add an optional
feature to not round the threshold up to the nearest second.  This was
causing some jumping back and forth from small to large, and the storm's
eye tracker wanted float thresholds and not the value rounded up to ints.