Add the infotext extension for timeline files
Example:
infotext "event name" before 2 "other text"

This would show "other text" as an info popup text on
the screen 2 seconds before the "event name" timer would
expire in the timeline.

infotexts are shown even for events that are hidden from the
timeline with hideall, under the assumption that the user
has requested this notification even if they don't care for
watching a timer go by.