Allow up to 2 lines of text popups of each type
This allows 2 info/alert/alarms to be visible at a time. If a third
becomes visible, it drops the oldest one.

Also some improvements to the CSS for raidboss to simplify things by
not repeating ourselves, and make each element of info/alert/alarm
text to be individually styled by giving them a class name.

We also improved the style of timeline fake move/resize bars to
better match the real bars.