Add the ability for the user to disable a trigger by its id
This gives a unique fight-and-ability-name-based id to each trigger that the
user should be able to disable, and adds an Options.DisabledTriggers field
that they can set with the ids to be disabled. Disabled triggers are still
run and still change the state of the |data| member, but will not appear
on screen or play any sounds.