Read rdm and war job specific data with the same helper function/signature
We discovered that rdm and warrior (and presumably other jobs) all store their job
specific data in the same place, and the code signatures were going to the same pointer.
So consolidate down to a single signature and a single method to read the job data, then
parse it out in job-specific code.