Add support for "elapsed" text in timerbar
"elapsed" here is the elapsed time, as compared to "remain".  "elapsed" is
chosen to avoid confusion with "value" which is often used as a
synonym for "remain" both internally and in the api.

Also clean up some existing text code that was inconsistently checking for
"value" when it really meant "remain".