Improve behaviour and accuracy of timerbar, and add tests
Now the timerbar uses a timestamp to measure progress so that it does
not slip if the timer callback is late. Also, setting the value after
setting the duration will advance the bar appropriately.

Also use RAF to animate the bar instead of a 100ms timer.

Adds a timerbar_test.html that verifies all these things.