Check the latest github release and warn if its newer
This grabs the page at http://github.com/quisquous/cactbot/releases/latest
and finds the version of the last release with regexs. If the version is
larger than the current assembly version (found in AssemblyInfo.cs) then
it shows a dialog pointing to the same URL. Once the user dismisses the
dialog it won't be shown again for the same version, but an error log is
printed instead on startup.