Fix o3s library phase triggers
Library phase stack/prey markers mess up the counting for spellblade holy,
which was counting sets of 4.  (It would also call out somebody to stack
on, which was also not super useful.)

Add a phase counter to differentiate which of these spellblade holies
it is.  Add a separate trigger for library phase markers.