ucu: Fix timeline syncs
Generate syncs apparently happen at the *end* of the castbar, so
add hidden syncs for each of these and adjust times accordingly.
This is what was causing the timeline to be off by three seconds
in phase 2 and phase 3.

Also, fix an overzealous sync on generate that caused Twintania P3 to
become incorrect by looping in the wrong spot.  Instead, just funroll
that loop and add the enrage at the end.