Support commas as decimal markers in floating point numbers
FFXIV plugin will put floating point numbers with commas for decimals in locales
where that is appropriate. So we have to parse numbers that way. However js
parseFloat() does not support that, so we have to convert comma to period before
calling parseFloat.

This adds a \y{Float} regex extension to patch floating point numbers across
locales. And a Regexes.ParseLocaleFloat() that can turn those matches into
a floating point number type.

The ParseLocaleFloat() is exposed to triggers on the data object as
data.ParseLocaleFloat() for triggers that want to match numbers, and this
changes all [0-9.] style matchers in the code and in triggers to use \y{Float}.

This should address the problems in bug #137.