Move timeline loading control to the trigger file
This means the zone extension is removed from the timeline file, and 2
new fields are added to the triggers definition:

timelineFile: <filename string>
This names a file in the timelines/ directory to be used when the
zoneRegex for the trigger set matches.

timelines: <multiline string>
This is additional text to append to the file (if there's a file) and used
as part of the timeline for that zone. Has the same format as the file
itself.

This also moves all the timeline extensions from the timeline .txt files
over to the triggers |timeline| fields, which allows the files themselves
to be shared with other timeline plugins that break in the presence of
unknown commands.