Fix parsing of comments and print errors when unable to parse a timeline
The parser was dropping anything after # regardless of context which is wrong. Change the
parser to instead only read commands that aren't after a # but accept still # in the middle
of a command such as sync/Foo #1/.

Any text left over at the end of parsing a line is not parsed, so print an error and don't
use that line when it occurs (other than whitespace/comments).