Make trigger-based timeline more functional
Previously, triggers could insert additional lines into timelines, and
this was primarily used to handle extensions to ACT timelines that the
base version didn't support (e.g. infoText).  These lines were only
able to be added as a single string.

This makes it possible to add arrays of strings (which maybe look nicer)
or functions (that take the same data object that triggers do).  This
makes it possible to conditionally hide or alert from timelines based
on job.

See: ui/raidboss/data/triggers/test.js for a cute example.