jobs: basic monk ui
Resource box (right of bars):
  chakra count, colored red when full

Timer bars (beneath hp/tp):
  one bar for greased lightning (different colors depending on stacks)
  one bar for form combo timer (different color for pb)

Timer boxes (below):
  blunt resistance
  twin snakes
  demolish

Currently, blunt resistance triggers on every mob and every application,
so will not track well with multiple mobs, but will help with multiple
people and one mob.

Demolish is also not mob-aware, but is tied to you, at least.

It'd be nice to make timer boxes have a threshold to get bigger when it's
known that it'll fall off before the next rotation.  However, monk gcd is
all over the place between grease and riddles, and so just leaving the
boxes to display seconds and threshold=0 for now.