jobs: fix monk ui bugs
(1) tornado kick doesn't reset greased lightning timer in memory, so set
remaining timer to zero if you dropped all your stacks.

(2) the filtering logic to only update greased lightning time if it
was larger (aka refresh) was overzealous and forgot to keep track of
when time got smaller so was only comparing refresh times.  This
caused refreshes to get dropped on the floor sometimes.

(3) the timer was sometimes jumping back to full, because the code was
setting duration after value, which implicitly resets the timer bar.