oopsy: custom mistake triggers
Support per zone trigger files, like raidboss.

Supported trigger entries (in order of evaluation)
* regex (matches any line)
* damageRegex (convenience, matches only ability name for attacks)
* condition (if false, then this line is ignored)
* runOnce (see below)
* delaySeconds (wait this many seconds before doing something)
* fooText (where foo is pull/warn/no/death/wipe, emits that mistake)
* run (don't emit, but run some code)

Anything that takes a function takes three parameters (event, data,
matches).  For damageRegex, event has a large number of fields on it.
For other regex types, it just has the line (in case you wanted it).

re: runOnce.

Many mistakes require evaluating multiple trigger lines, for instance
making sure N people are stacked for a fireball, that you don't take
two o4s lasers by standing in the middle, or that you don't double
tap somebody with your o1s lightning puddle.  This is kind of a
pain to do in raidboss, and usually requires special variables etc.
If runOnce is true, then all matched lines for this trigger are
accumulated and then after delaySeconds, are all passed to each
fooText function and the run function.

This means that for runOnce triggers, the event parameter for
fooText and run will be an array and not just a single event.
(Note: condition still only gets the one event, because it could
prevent any single event from triggering).