Differentiate game vs act inCombat
ACT's "in combat" lasts for as long as ACT's timeout window is configured
for.  C# also detects wipes (where you come back to life without weakness)
and will stop ACT in those cases.  ACT also considers you in combat if
somebody else nearby is hitting a striking dummy even if they're not in
your party.  So, it can be a little bit wonky.

FFXIV itself also has a notion of "in combat" (e.g. when you can/can't
teleport, when anybody else in your party is attacking something) which
ends much much sooner than ACT's timeout window.

This adds a memory signature and event for "in game combat" vs "in act
combat" and modifies modules to use this new data.  The goal here is
to remove OnBossFightStart from C#.

Note: Steps of Faith is an edge case here, as it is for wipe detection
(with its return mechanics), and it appears that "in game combat" is
false when you are dead there.  This is not true anywhere else.