raidboss: reset data when starting combat too
This is the root cause of the ucu dragon error,error,error message.
Sometimes, combat would end, but there'd still be an attack from the
holy dragon, for example.  This would record the dragon's position.
Then the next run would start.  That dragon has already been seen so
would never be recorded.  Once the other four were seen, then it would
try to calculate dragon positions.  If you were lucky, the holy
dragon from the previous run would be in the same spot as another dragon
and you'd get errors.  If you were unlucky, then the holy dragon
would be thought to be in a different spot and you'd get some wrong
dives.

This should fix that once and for all by resetting data when combat
starts as well.