Allow timelines to be manually stopped.
Currently, timelines and triggers stop automatically when the game is no
longer in combat.  This appears to happen in O5S during the ghost add
section while the main train goes on a brief vacation.

This adds a band-aid option for trigger files to specify that their zone
needs to be manually stopped.  A trigger can then call data.StopCombat(),
which will stop both timelines and triggers.

Wasn't able to repro the O5S bug in O5N, but manual testing of this flag
and of the settings inside of O5N make it seem plausible that this will
fix the bug in O5S.

Closes #33.