Add DispatchToJS lock to prevent races
These data races have always been there, but were only going to be
exposed in PullCounter, which called an OverlayMessage that dispatched
an event, which could collide with game events dispatched from the
fast thread.

These races result in bogus Javascript to execute, such as such as
    document.dispatchEvent(new CustomEvent('onSendSaveData', { detail: {document.dispatchEvent(new CustomEvent('onSendUserConfigLocation', { detail: "data":" etc

...because both of these events are trying to be sent simultaneously,
and the string builder still has text from the previous event in
progress.  This leads to anonymous Javascript errors like "Unexpected
token )", which are awfully confusing.  It also sometimes but not
always leads to JSON serialization C# errors.

(This could explain why PullCounter never started when loaded
remotely as well.  onSendSaveData could be clobbering other events.)

This should fix races that will become a lot more common in pull #64.