Simplify user location messages
If remote user directories aren't going to check for existence, then
there's no reason to do it in the local case either.  This cleans up
needing to send the overlay name along with the message and just
allows the remote side to send the user directory alone.

This also removes the checks for css and js, which allows for
overriding css without overriding js in the case of local files.