Wait for user config before initializing overlay
This is the followup to 476ea6d which changes overlays to wait on the
onOverlayInitialize message before creating any of their objects,
reading options, or hooking up to DOM elements.  This should stop
initialization races ONCE AND FOR ALL.

This also reverts most of 65cf263 which half-heartedly tried to fix this.