Adding O8S timeline syncs (#66)
* raidboss: Adding O8S timeline syncs

The diff may not be super useful, so I'll summarize the changes (beyond
simply adding syncs). Preserved the existing naming for the most part.

- Added the Thunder/Fire hits of Celestriad
- Fixed some slight +/- 1s timing issues
- Added all non-starstrafe Soaks (The Path of Light, renamed for
clarity), since they were included, albeit inconsistently, later on in
the fight. Having specific timing for them is useful for dodging into
it.
- Added All Things Ending to Past/Future End during Starstrafe. The ATE
cast on the normal ones is fast enough (3s) that it's not worth timing
it, but there's a bigger gap in that one (8s).
- Also added Meteor hit (the stack) during Starstrafe
- Added some missing Ultimas (1328, 1358, 1543)
- Added a missing Ultimate Embraces (1364)
- Changed tether times to activation rather than appearance

* raidboss: O8S timeline minor fixes

Forgot to save these.

* raidboss: O8S end of fight syncs