Fix null config directory preventing overlays from starting
The Uri constructor apparently throws exceptions (oops) and this was
causing the overlay to explode and not send the init message.  This
would happen if you hadn't set the user config directory.  I had
tested this earlier, but not after the most recent user file sending
changes that broke it.

SORRY T_T

Fixes #76.