raidboss: Cleaning up timelines
Fixed some timing issues in Ridorana and added the add phase in Tsukuyomi. Weirdly, line 17 doesn't work even though I'm sure that line is present in the logs, Not sure what's up with that, but it only causes Concentrativity to show the wrong timing and it's pretty clearly visible. Might want to just hide that for now though.