Cleaning up timelines (#129)
Fixed some timing issues in Ridorana and added the add phase in Tsukuyomi. 

Upon more playing, I discovered that the last phase begins after the second Selenomancy phase regardless of boss HP, and also that the nightfall during that phase is always gun the first time and always spear the second time.