Fix Ridorana math
Looking for current hp + boost is racy.  From a local test, t=0
and t=0.5 after subtract do not include the boost (even if
standing in it), but t>=1 do include the boost.

Therefore, attempt to record the current hp at t=0 and t=0.5
and find a single digit hp to do the mechanic with, and use this
the whole time.

This is probably going to be wrong in some cases if folks are
lagging, but there's not much to do about it.