Remove redundant tts
tts now defaults to alarm > alert > info text if it is available.
This means that triggers that duplicate one of these texts for
tts no longer need to and can just inherit the tts text by
default.

However, if *any* language needs a different tts, then tts needs to be
specified for all of them.