Create ja.js
Japanese translation. Some comments:

Under this.areaSealRegex, there is a blank space right before 秒. I did this because there was a blank space in the other translations, but Japanese client appends 秒 after the number (indicating seconds). this.countdownStartRegex also has the 秒 marker after the end of 
(\y{Float}) for reasons I described.