raidboss: o9s trigger cleanup
This adjusts some triggers from commit bad1b99.

Added todos.

Made lat/long/edict tell you what to do, not what it does.  This is maybe
less clear when you have Primordial Crurst, but we can clean that up later
so that debuffs say "Die on sides" instead of "Sides -> Front/Back".

Adjusted Fiendish Orbs id based on fflogs (hopefully not breaking it).
It seems like it should be 317C and not the other.

Added ids to various triggers from reading fflogs.

Added a bunch of role conditions for various triggers to make them less
noisy.

Removed aoe callout, as that's pretty noisy and healers often prep for
that way sooner than when the ability starts.