raidboss: o12n timeline
I am not 100% sure, but I think that the first phase loops until a low%
and then jumps into passage of arms.  I think also that the last phase
picks shield or blade to start with, and then alternates.  However, I
couldn't find a slow enough parse on fflogs to verify that it didn't
ever repeat one.