Adding my personal changes. Removing sounds/Overwatch/D.Va_-_Game_on.ogg (I use my own version, I'll upload this and fix the syntax later.)
Also changing the Octet trigger for Twintania to something not terrible. YOU stack for Twin is silly. YOU BAIT Twin is much better.

Finally, we add a raidboss2 and seperate the timeline and triggers to make it easier to customize on the UI. Obviously this will mean you need to add a new overlay, but I think its a small price to pay.