Fix race condition in startup
This setTimeout could potentially race with the initial messages
sent from the plugin.  As the callback() call inside the timeout
is what sets up most of the objects to receive the messages, if
the c# plugin starts delivering before the setTimeout runs,
then there will be errors.

This race can easily be seen by making the setTimeout have a
3000 wait instead of a 0 wait.

To fix this, make the javascript run synchronously to stop this.

Fixes #298.