fisher: handle plural fish names and german plurals
In German, various fish would sometimes be returned as plural,
such as Amoniten for a single Ammonite.  Additionally, words
with [a] in them can have different endings depending on
context.  So, add support for all plural fish name (german
or otherwise) as well as all combinations of [a] endings
so that all fish can match.

To support this, this allows entries in the fishing database to be
either strings or lists.  The first entry in the list is the primary
entry and will always be returned, but other entries in the list can
also match for name lookups.

The previous places_cast values have been folded into places,
as this is now supported behavior.

This also will fix double hook returning multiple fish.