raidboss: add fake loop lookahead for e1s
This is an update to d98c6dd.

This is just personal preference, but I like to at least add 30+ seconds
of the abilities before looping back.  I use the timeline_adjust script
to make sure that the abilities are exactly the same times just shifted
forward.  (This is maybe a TODO for a loop command that would do this
automatically that could replace the jump command.)

It's also a personal preference to jump from one ability to the same
ability previously, rather than into empty space.  It's just more
readable for me.