raidboss: change dps role to be 'dps' (breaking)
Previously, data.role for dps, was dps-caster or dps-melee or
dps-ranged.  This was super awkward and was really only used
for role-related abilities like addle, etc.  This change makes
the dps role much more understandably just 'dps'.

Because this touches a bunch of triggers that related to apoc,
addle, and feint, also centralize those calls.

Also, remove everything related to apoc and most of the o4s
triggers related to addle.

This also moves all of these functions into common.js so
raidboss and the emulator can share them for their data obj.