fisher: handle The Red Chalice properly (#539)
The Red Chalice appears as the name of the fishing zone in the
game data: https://xivapi.com/FishingSpot/226

However, when casting, this turns into "the Red Chalice" and gets
eaten by the optional "the" that appears in front of many English
fishing spots.

Keep the generated data the same, and handle this at runtime.