jobs: add more abstractions
Apply this to war first.

Currently, jobs is a lot of spaghetti code.  There's a lot of copy and
paste to set up html and style blocks.  There's functions all over the
place to handle combos and job updates.  It's very tricky for somebody
to come in add a new job.

Also, every log line gets read by a ton of different regexes, which is
some of what I suspect is going on with delays on the timers.

Ideally, this will consolidate most of the job logic for each job
into a single place.  It should be then easy enough to add functions
when effects/abilities/gauge info changes and manipulate those boxes.

As a followup, most of the css will also be switched to put multiple
resource/proc boxes in a reasonable place such that per-id css
doesn't have to be used to lay stuff out.