jobs: add MP tick timer for BLM
BLM is a job that relies heavily on exact MP tick timings, since getting
a lucky MP tick early on in the umbral ice phase allows BLMs to skip
needing a second filler spell for mana.

This commit adds support for tracking the MP tick timer. To detect MP
tick events, we hard-code the known constants for MP tick numbers (these
were known even before SHB, which is why I'm hard-coding them as a
percentage of max MP rather than using exact values, even though max MP
is now fixed at 10000).

To continue tracking the MP tick even when mana ticks are disabled (such
as in astral fire) or mana is high enough that excess mana from ticks is
cut off, we extend the timer-bar class to support seamlessly looping the
timer.

The user is able to control the display of this MP ticker as well as
hide it when out of combat, which is the default behavior.

We also color the MP ticker color based on the umbral state, to more
clearly communicate to BLMs whether they will gain an MP tick or not.