jobs: add buff countdowns (#550)
Now when buffs are used, if there's a specified cooldown, it will show a
countdown for when that buff is ready.  When it's entirely ready it will
show without a timer so you can be extra sad when that pug monk keeps
desyncing their brotherhood from your tank buffs.

Because the countdown starts when the effect occurs (and not when the
ability is used) it will lag slightly behind by about half a second.
I might consider just cheating the time forward by some heuristic to
make it feel better if this is jarring.

Times are also a little off because this uses setTimeout instead of
storing times.  Recalculating better remaining time is left for the
future as well.

This is a first cut, feedback welcome.  I think the UI location and
differentiation from active buffs needs work, especially.  However, I
wanted an initial version to play with and show other people.