This eliminiates a race condition during raidboss initialization.
If we attach the overlay listeners before gPopupText has been set,
we risk triggering a "Cannot read property '...' of undefined"
error. If we attach them in the constructor instead, we can be sure
that the object exists.
Since only one instance of PopupText is ever created, we don't have
to worry about detaching the overlay listeners.

This should fix the issue reported in
https://github.com/quisquous/cactbot/issues/660#issuecomment-542819459