plugin: remove FFXIV_ACT_Plugin references
Because of how C# loading works, if people store this plugin in the
root directory of ACT, if cactbot references it directly, then it may end
up getting loaded into a different context.

Intsead, remove references to the plugin, and in the future reference
the broken out dlls in the sdk.

See also:
https://github.com/quisquous/cactbot/pull/577#issuecomment-531517955