lint: Add ESLint Recommended Preset (#681)
Add the ESLint "recommended" preset which includes a few nice checks,
most notably the checking of multiple spaces within a regular
expression, which was updated in #677. Additional checking for
JavaScript logic and syntax is included as well.

Compresses, reorders, and updates the package.json rules section to now
exclude duplicate rules and to explicitly say "error" and "off", over 2
and 0, respectively.