raidboss: allow for separate timelines and alerts
This adds two new html files, which redirect to the original with
some additional query parameters.

It's more cpu work to separate these, because both the alerts and
the timeline version have to run alerts and timelines, but just
hide them.

However, a lot of people have asked for this, so hopefully it will
be helpful to folks who don't want to have to sort out CSS to
move these around.

Note: these new html files do not work with ACT websocket, sorry.
HTTP redirects don't preserve these query parameters and it's
awkward to try to do this in other ways.

If you have to have ACT websocket + alerts or timeline only, you
need to manually append `&alerts=0` or `&timeline=0` to your url.

Note: this is also only for the ngld overlay, because it supports
URLSearchParams, which Chrome M43 does not.