add party tracker class
Both oopsy and raidboss now grow a `data.party` member which is an
instance of the PartyTracker class in `party.js`.  That file exposes
a number of helper functions about the party.

This will allow for triggers in both to depend on party member
information.

This will only be filled in when using ngld's overlayplugin.
If using hibiya's version, `data.party` will exist, but all functions
will return false or empty arrays.

Fixes #694.