Add canned regex helpers (#740)
Adds a number of regex helper functions that all take a dictionary of optional fields. Fields can be filled in optionally with strings/regular expressions/arrays of multiple values. If field values are unspecified, it will match anything.

Each dictionary of fields can also specify capture as a boolean variable. If not specified, capture is true. If capture is true, then everything in the regex will be captured as a named group (with the same name as the field that would have specified it).

If a regex has capture groups, the `matches` parameter for a trigger will be the `groups` property of the result rather than the result of the regex.  See the o5n example.

Examples:

Regexes.losesEffect({effect: '(?:Doom|Forked Lightning)'})
Regexes.startsUsing({id: ['1C7F', '1C90'], name: 'Deathgaze Hollow', ability: 'Void Death'})
Regexes.ability({name: 'Caduceus', id: '4B8', ability: 'Hood Swing'})

Fixes #543.