raidboss: TEA timeline fixups
This is an addendum to #760 after running test_timeline.

Commenting out the first ability usage in pairs of three where they
are all close together to be extra careful not to sync incorrectly.

Chakrams and Final Sentence (maybe?) only appear in the log
when somebody gets hit by them, I'm guessing, so commenting those out.