oopsy: more TEA warnings
There's a giant pile of TODOs here for things that I was less sure about,
so please send pull requests if you'd like.

Also, it seems pretty clear that maybe oopsy needs a built in
"warn if this hits more than two people" category, just based on the
number of manual triggers that require that.