Make `response` a possible trigger entry
Because the spread ... doesn't work in an eval context,
allow for a trigger entry point called `response` that can return
an object that contains text and tts properties.

(This will also allow for triggers that want to return multiple
texts to do so in one place and not save the text it wants to print
in data or duplicate code).

Also, update Conditions to be functions to match Regexes and Responses.