plugin: fix reloading when user files changed
When the DLL moved as a part of the publishing folder structure change for
ngld, this broke automatically finding the user directory.

Also, Directory.Exists does not work properly with mixed slashes, so use
the Uri constructor to canonicalize the uri for use in determining if it
exists.

Finally, actually hook up the event to reload the overlay when files
change.