print the names of all user/ files loaded
This is maybe noisy, especially given that cactbot overlays have a
weird startup where they get reloaded a few times as other overlays
finish.

However, given the amount of debugging of "is my user file begin loaded"
and "why is my tts not working", it seems worthwhile to do this.