Update monster position (#928)
Quick change to update monster position when (re-)detected, mainly for B ranks but works for all monsters.

This fixes the situation where (example):
1. You're scouting and gets in range of a B rank.
2. Radar detects B rank and adds it to the list.
3. You fly out of range as you continue scouting.
4. B rank gets killed while you're out of range, causing it to respawn somewhere else.
5. You get in range of said monster's new position.
6. Radar doesn't update the position because the old one is "still alive".

This small change makes it so the radar update the B rank's position in example above. Updating the position in realtime would be suitable but I don't know enough to code it. This should however be enough as hunts don't move too far anyway.

Adds ability to tell if its closer and only update if needed. Additionally, only play sound if its far enough from the original position (hardcoded to 100 for now)