regexes: add target field to ability regex
I think this is is just so common that to not have it will cause ongoing
errors in the future.

This seems to be the cause of the fireball/thunderstruck errors in the
ucob triggers.  Because matches.target will always be undefined, it
will never match that player.

It's possible that maybe we need some hacky unit tests to see what
properties are being used on matches, and making sure that those are
valid fields based on the regex used.  There's not a clean way to do this,
but it seems possible that we could catch a lot of errors.

Fixes #966.