config: run parseInt/parseFloat on any integer/float settings
Currently everything is stored as strings, because it made writing the
code easier when folks are inputting transitory inputs that may not be
valid (like starting `.3` with the `.`)

However, a lot of other code expects these to be integers and not strings,
and this causes issues.  See #1000 for one example.